# [tnx] Read-only Neuron cache workaround for SM