--- a/engines/python/setup/djl_python/neuron_utils/model_loader.py
+++ b/engines/python/setup/djl_python/neuron_utils/model_loader.py
@@ -33,6 +33,33 @@ class ModelLoader(ABC):
     def __init__(self, *args, **kwargs) -> None:
         self.config = kwargs.get("config")
         self.model_config = kwargs.get("model_config", None)
+        self.move_read_only_neuron_cache()
+
+    def move_read_only_neuron_cache(self) -> None:
+        """
+        If the currently set Neuron cache directory is read-only,
+        change the Neuron cache directory to the default: /var/tmp/neuron-compile-cache.
+        Copies the graphs from the original location to the default location.
+
+        This is a workaround for passing the Neuron cache with the model on SM.
+        Enables reading a Neuron cache located in the read-only dir /opt/ml/model.
+        """
+        cache_dir = os.environ.get("NEURON_COMPILE_CACHE_URL")
+        default_cache_dir = "/var/tmp/neuron-compile-cache"
+        if cache_dir:
+            cache_dir = os.path.abspath(cache_dir)
+            if not os.access(cache_dir, os.W_OK):
+                logging.info(
+                    f"Neuron cache directory is set to an unwriteable location: {cache_dir}"
+                )
+                start = time.perf_counter()
+                shutil.copytree(cache_dir, default_cache_dir)
+                os.environ["NEURON_COMPILE_CACHE_URL"] = default_cache_dir
+                duration = time.perf_counter() - start
+                logging.info(
+                    f"Copied neuron cache to the default location: {default_cache_dir}. "
+                    "Using this directory as the Neuron cache."
+                    f"\nCopying took: {duration} seconds")
 
     def init_load_path(self) -> str:
         """